Move rest of checkout to the cart app
