+++ saleor/cart/checkout/views/__init__.py
@@ -1,18 +1,37 @@
 from django.shortcuts import redirect
 from django.template.response import TemplateResponse
 
-from ....checkout.core import load_checkout
-from ....checkout.views import add_voucher_form, validate_voucher
-from ....checkout.views.validators import (
-    validate_cart, validate_is_shipping_required, validate_shipping_address,
-    validate_shipping_method)
+from ....account.forms import LoginForm
+from ..core import load_checkout
 from ..forms import CartShippingMethodForm
 from ..utils import get_checkout_data
+from .discount import add_voucher_form, validate_voucher
 from .shipping import (
     anonymous_user_shipping_address_view, user_shipping_address_view)
 from .summary import (
     anonymous_summary_without_shipping, summary_with_shipping_view,
     summary_without_shipping)
+from .validators import (
+    validate_cart, validate_is_shipping_required, validate_shipping_address,
+    validate_shipping_method)
+
+
+@load_checkout
+@validate_cart
+def login(request, cart, checkout):
+    """Allow the user to log in prior to checkout."""
+    if request.user.is_authenticated:
+        return redirect('cart:checkout-index')
+    form = LoginForm()
+    return TemplateResponse(request, 'checkout/login.html', {'form': form})
+
+
+@load_checkout
+@validate_cart
+@validate_is_shipping_required
+def index_view(request, cart, checkout):
+    """Redirect to the initial step of checkout."""
+    return redirect('cart:checkout-shipping-address')
 
 
 @load_checkout
@@ -42,7 +61,7 @@ def shipping_method_view(request, cart, checkout):
 
     if form.is_valid():
         form.save()
-        return redirect('checkout:summary')
+        return redirect('cart:checkout-summary')
 
     ctx = get_checkout_data(cart, request.discounts, taxes)
     ctx.update({
