+++ templates/cart/index.html
@@ -77,7 +77,7 @@
       </div>
       <div class="row">
         <div class="col-md-12">
-          <a href="{% url "checkout:login" %}" class="btn primary float-right cart__submit">
+          <a href="{% url "cart:checkout-login" %}" class="btn primary float-right cart__submit">
             {% trans "Checkout" context "Cart primary action" %}
           </a>
         </div>
