+++ templates/checkout/shipping_method.html
@@ -6,7 +6,7 @@
 {% block forms %}
   <h3>{% trans "Shipping address" context "Checkout shipping address title" %}</h3>
   {% format_address cart.shipping_address %}
-  <p><a href="{% url 'checkout:shipping-address' %}">{% trans "Select other address" %}</a></p>
+  <p><a href="{% url 'cart:checkout-shipping-address' %}">{% trans "Select other address" %}</a></p>
   <hr>
   <form method="post" novalidate>
     {% csrf_token %}
